refactor(api): extract admin user routes to admin controller
